requirements/requirements-base.txt
python-intercom>=3.1.0
celery>=4.0.2
django-jsonfield>=1.0.1
